requirements.txt
aiogram==2.1
gspread
google-auth
google-auth-oauthlib
google-api-python-client
psycopg2-binary
openai
aiohttp
pdfkit
gunicorn
